requirements.txt
nltk==3.5
nonebot==1.6.0
jieba==0.42.1
aiohttp==3.6.2
aiocqhttp==1.2.5
requests==2.23.0
numpy==1.18.4
pyahocorasick==1.4.0
httpx==0.12.1
ahocorasick==0.9
beautifulsoup4==4.9.0
dataclasses==0.7
Pillow==7.1.2
scikit_learn==0.22.2.post1
